// precotex_angular/src/app/models/Estampado/ProgramaEmpastado.ts
export class ProgramaEmpastado{
    Id_Programa: string = '';
    Cod_Cliente: string = '';
    Nom_Cliente: string = '';
    Cod_Ordtra: string = '';
    Num_Secuencia: string = '';
    Cod_Diseno: string = '';
    Cod_Comb: string = '';
    Cod_Combinacion: string = '';
    Cod_Tela: string = '';
    Des_Tela_Abr: string = '';
    Cod_Color: string = '';
    Des_Color: string = '';
    Kgs_Progr: string = '';
    Engome: string = '';
    Cod_Ordtra_Transf: string = '';
    Cod_Und_Negoc: string = '';
    Des_Und_Negoc: string = '';
    Observaciones: string = '';
    Nom_Archivo: string = '';
}

export class VersionPrograma{
    Version: string = '';
}

export class TipPasta{
    Cod_Tip_Pasta: string = '';
    Nom_Tip_Pasta: string = ''; 
}
    